Move screen-size listener to custom hook
